+++ src/function/handler.ts
@@ -1,23 +1,26 @@
-import type { Context } from './context.js'
-import type { Event } from './event.js'
-import type { Response, BuilderResponse, StreamingResponse } from './response.js'
+import type { HandlerContext } from './handler_context.js'
+import type { HandlerEvent } from './handler_event.js'
+import type { HandlerResponse, BuilderResponse, StreamingResponse } from './handler_response.js'
 
-export interface HandlerCallback<ResponseType extends Response = Response> {
+export interface HandlerCallback<ResponseType extends HandlerResponse = HandlerResponse> {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   (error: any, response: ResponseType): void
 }
 
-export interface BaseHandler<ResponseType extends Response = Response, C extends Context = Context> {
-  (event: Event, context: C, callback?: HandlerCallback<ResponseType>): void | Promise<ResponseType>
+export interface BaseHandler<
+  ResponseType extends HandlerResponse = HandlerResponse,
+  C extends HandlerContext = HandlerContext,
+> {
+  (event: HandlerEvent, context: C, callback?: HandlerCallback<ResponseType>): void | Promise<ResponseType>
 }
 
-export interface BackgroundHandler<C extends Context = Context> {
-  (event: Event, context: C): void | Promise<void>
+export interface BackgroundHandler<C extends HandlerContext = HandlerContext> {
+  (event: HandlerEvent, context: C): void | Promise<void>
 }
 
-export type Handler = BaseHandler<Response, Context>
-export type BuilderHandler = BaseHandler<BuilderResponse, Context>
+export type Handler = BaseHandler<HandlerResponse, HandlerContext>
+export type BuilderHandler = BaseHandler<BuilderResponse, HandlerContext>
 
 export interface StreamingHandler {
-  (event: Event, context: Context): Promise<StreamingResponse>
+  (event: HandlerEvent, context: HandlerContext): Promise<StreamingResponse>
 }
